the code in the dashboard has been refactored to give an enhanced developer UX
