+++ client/src/components/WorkoutList.tsx
@@ -1,0 +1,98 @@
+import { Plus } from "lucide-react";
+import { useNavigate } from "react-router-dom";
+import axios from "axios";
+import Workout from "./Workout";
+
+function WorkoutList({ username, workouts, setWorkouts, setWorkoutUpdateId }) {
+  const navigate = useNavigate();
+  const token = localStorage.getItem("token");
+
+  const signOutButton = async () => {
+    try {
+      const response = await axios.delete("http://127.0.0.1:8080/auth/logout", {
+        headers: {
+          Authorization: `Bearer ${token}`,
+        },
+      });
+      console.log(response.data.message);
+    } catch (error) {
+      console.error(
+        "Failed to sign out user:",
+        // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+        // @ts-expect-error
+        error.response?.data?.msg || error.message
+      );
+    }
+    localStorage.clear();
+    // TODO: fix this
+    navigate(0);
+    console.log("The local storage has been cleared");
+  };
+
+  const onDelete = async (workoutId: number) => {
+    const url = `http://127.0.0.1:8080/${workoutId}/delete`;
+    const headers = {
+      Authorization: `Bearer ${token}`,
+    };
+    try {
+      const response = await axios({
+        method: "post",
+        url: url,
+        headers: headers,
+      });
+      setWorkouts(response.data.workouts);
+    } catch (error) {
+      console.error(
+        "Failed to delete workout log:",
+        // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+        // @ts-expect-error
+        error.response?.data?.msg || error.message
+      );
+    }
+  };
+
+  return (
+    <>
+      <div>
+        <div className="flex items-center justify-center">
+          <h2 className="text-2xl font-semibold text-gray-800 dark:text-white">
+            Welcome {username}!
+          </h2>
+        </div>
+      </div>
+      <div className="flex justify-center items-center">
+        <div className="w-full max-w-xl">
+          <div className="flex items-start justify-between mb-2">
+            <button
+              className="px-4 py-2 mb-2 bg-blue-500 text-white rounded-md flex items-center gap-2 hover:bg-blue-600 cursor-pointer"
+              onClick={() => navigate("/workout-log")}
+            >
+              <Plus className="w-5 h-5" />
+              Add Workout
+            </button>
+
+            <button
+              className="bg-red-500 text-white flex items-center gap-2 px-8 py-1 rounded-md hover:bg-red-600 cursor-pointer justify-around"
+              onClick={signOutButton}
+            >
+              Sign out
+            </button>
+          </div>
+
+          <div className="container bg-gray-100 p-6 rounded-xl">
+            {workouts.map((workout, index) => (
+              <Workout
+                workout={workout}
+                index={index}
+                setWorkoutUpdateId={setWorkoutUpdateId}
+                onDelete={onDelete}
+              />
+            ))}
+          </div>
+        </div>
+      </div>
+    </>
+  );
+}
+
+export default WorkoutList;
